Break out metadata management from `pkg/cmd/pulumi/util`
`pkg/cmd/pulumi/util.go` has a lot of diverse functionality within it.
This commit starts the process of breaking this out into subpackages so
that things can be better namespaced and encapsulated, factoring out
operation metadata into a new `metadata` package.
