+++ pkg/cmd/pulumi/up.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -25,6 +25,7 @@ import (
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
+	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/metadata"
 	"github.com/pulumi/pulumi/pkg/v3/engine"
 	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
 	"github.com/pulumi/pulumi/pkg/v3/resource/stack"
@@ -118,7 +119,7 @@ func newUpCmd() *cobra.Command {
 			return fmt.Errorf("getting stack configuration: %w", err)
 		}
 
-		m, err := getUpdateMetadata(message, root, execKind, execAgent, planFilePath != "", cfg, cmd.Flags())
+		m, err := metadata.GetUpdateMetadata(message, root, execKind, execAgent, planFilePath != "", cfg, cmd.Flags())
 		if err != nil {
 			return fmt.Errorf("gathering environment metadata: %w", err)
 		}
@@ -372,7 +373,7 @@ func newUpCmd() *cobra.Command {
 			return fmt.Errorf("getting stack configuration: %w", err)
 		}
 
-		m, err := getUpdateMetadata(message, root, execKind, execAgent, planFilePath != "", cfg, cmd.Flags())
+		m, err := metadata.GetUpdateMetadata(message, root, execKind, execAgent, planFilePath != "", cfg, cmd.Flags())
 		if err != nil {
 			return fmt.Errorf("gathering environment metadata: %w", err)
 		}
