+++ pkg/cmd/pulumi/watch.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2019, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import (
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
+	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/metadata"
 	"github.com/pulumi/pulumi/pkg/v3/engine"
 	"github.com/pulumi/pulumi/pkg/v3/resource/stack"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
@@ -113,7 +114,7 @@ func newWatchCmd() *cobra.Command {
 				return fmt.Errorf("getting stack configuration: %w", err)
 			}
 
-			m, err := getUpdateMetadata(message, root, execKind, "" /* execAgent */, false, cfg, cmd.Flags())
+			m, err := metadata.GetUpdateMetadata(message, root, execKind, "" /* execAgent */, false, cfg, cmd.Flags())
 			if err != nil {
 				return fmt.Errorf("gathering environment metadata: %w", err)
 			}
